Change dependence between costs and land, etc.
This commit includes three major updates:
(1) First, we changed the dependence structure between migration costs
and landholdings to ensure that (i) the parameters a0 and a1 would be
unconstrained, and (ii) that the conditional cost distributions would
have strictly increasing densities.
(2) Second, we fixed an error in the way that we were converting units
between model-implied displacement, which is per family, and observed
displacement, which is a head count. Previously we multipled by pop/fam
but the correct scaling is simply pop.
(3) Third, we changed the parameters at which we simulated displacement
data to reflect (1) and (2) above so that our simulation would yield
counts roughly on the scale of the real displacement data.