Changes to prevent 0 migration from non-zero V
We made two changes to address the problem that occurs when all
municipalities have zero predicted migration: the inner optimization
fails because the measurement error model requires variation in
predicted migration for identification.

(1) We tweaked our fixed point iteration slightly. The original version
returned the *previous* iterate rather than the *current* iterate. This
is wasteful and also meant that an iteration starting from zero would
terminate at zero provided that the first iteration yielded a function
value less than the tolerance (0.0001).

(2) It could still be the case that predicted migration is so small that
after adding the starting value for d_bar in the inner iteration,
underflow results in there being zero variation in the sum of d_bar and
dstar_friction. If this occurs, the inner optimization is not carried
out, as its parameters would not be identified. Instead, we simply
return the largest value representable on the machine in lieu of the
actual negative log-likelihood.