Change profile log-like for displacement
We found a closed form for nu and eta given rho and dbar. Using these
calculations, we changed our procedure for calculating the profile log
likelihood as a function of theta. The "inner" optimization is now over
rho and dbar, which lie in a bounded parameter space over which we have
analytical derivatives. We use the L-BFGS-B procedure from optim.

We also made some small housekeeping changes to the code that simulates
from the model so that it is more succinct and uses the same conventions
as our updated code for the profile log likelihood.