Changes to model of displacement
This commit makes two substantive changes to the structural model of
displacement:

(1) Rather than dividing the return r to land by mu, the mean
landholding in a municipality, we now divide it by the median
landholding in the municipality. Land distributions are very skewed, so
the mean is a very poor measure of centrality and (we suspect) was
causing changes in r to have very small effects on the integration
region.

(2) We previously thought that the first scale parameter of the cost
distribution had to be 2 or greater to ensure an increasing density.
(The second scale parameter is fixed at 1.) However this is incorrect:
it only needs to be greater than 1. Accordingly we changed all instances
of (2 + a0) to (1 + a0)

We also made some changes to the simulated data:

(3) After changing the way we scaled r in (1) above, we likewise changed
the true value of r in the simulation so that it made sense relative to
tau_ell etc.

(4) There was an Rda file called dstar.Rda hanging around in the
package. It's a holdover from some earlier testing code that is no
longer needed, so we deleted it.