Ensure that cum migration is nondecreasing
Due to the finite precision of our fixed point algorithm, it turned out
that at certain parameter values the sequence of cumulative migration
within a muncipality could slightly *decrease* (in say the 8th decimal
place) across periods in which there was no change in violence. We fixed
this by telling the code that calls the fixed point algorithm to take
the max of the current and past period displacement.