Add more comprehensive panel data
Our original raw panel dataset was Panel_D_V.dta. This only contained
information for municipalities for which we observed the land
distribution. But to construct average violence in neighboring
municipalities, we need to know the violence in places where we might
not have information on the land distribution. We went back and checked
the original STATA files and determined that Panel_D_V.dta was a subset
of the original, more comprehensive file that includes displacement
measures and violence data for 1051 municipalities. We have added this
more comprehensive dataset as are "new" raw panel dataset: it's called
panel-raw.dta. This changed required a small modification at the top of
process_raw_data.R to refer to the new file.

Before making these changes we checked that all observed values for
municipalities that were present in both Panel_D_V.dta and panel-raw.dta
did in fact coincide. These were identical.