Add expanded raw cross section data
Initially we included only data for muncipalities with observed land
distributions. Here we add a file with information for *all*
municipalities, regardless of whether land characteristics are observed.
Some testing code in this commit verifies that the data in the
Cross_Section_final_RAW_1076.dta matches that in cross-section-raw.dta
for any muncipalities that are in both files.