simplify shapefiles remove old functions
The shapefile CO_shape was very large making it very slow and tedious to
plot maps. We "downsampled" the polygons to preserve only 5% of the
original points. This gives us plenty of resolution. We also removed the
islands. Since we no longer need the old plotting functions, we removed
these as well.